# Refactor gaussian_blur in numpy to utilize scipy.signal.convolve2d