To improve performance, I optimized the convolution logic in gaussian_blur by replacing the NumPy-based implementation with scipy.signal.convolve2d. 
Below are the performance test results on my local machine:


- Before Optimization:
```
%%time
%load_ext autoreload
%autoreload 2

from keras.src.backend.numpy.image import gaussian_blur

images = np.random.uniform(size=(2, 800, 800, 3)).astype("float32")
kernel_size = (3, 3)
sigma = np.random.uniform(size=(2,)).astype("float32")

transformed_image = gaussian_blur(images, sigma=sigma)
```
CPU times: user 7.59 s, sys: 0 ns, total: 7.59 s
Wall time: 7.59 s

- After Optimization:
```
%%time
%load_ext autoreload
%autoreload 2

from keras.src.backend.numpy.image import gaussian_blur

images = np.random.uniform(size=(2, 800, 800, 3)).astype("float32")
kernel_size = (3, 3)
sigma = np.random.uniform(size=(2,)).astype("float32")

transformed_image = gaussian_blur(images, sigma=sigma)
```
CPU times: user 123 ms, sys: 0 ns, total: 123 ms
Wall time: 122 ms